Establish tracked-item mongoose model
Instead of using a config file to store the tracked items, switch to
use a mongo/mongoose model.  Pull the logic which scrapes the sites
from the controller to a library file.  Build a controller to expose
tracked items, and for now, expose an update API to update the prices.