Create an add tracked items page
Create a page that will allow a user to add a tracked item.  Once they
enter in the URI, clicking "populate" should fill in the name,
category, and current price.  Once that is shown, the user will be able
to submit the item to the tracked items list.  (This is yet to be
implemented.)