Implement fetch item details from UI
When a user enters in a URI in the add tracked item screen, they can
click on the "populate" button which issues a POST request to the
server with only the URI sent for a tracked item.  The server knows
this must mean the user wishes to gather the item details, and does so,
returning the populated information.  The UI then populates the
remainder of the screen with the details on the item.

From here the user can click Add to create the tracked item on the
server.  (This has yet to be implemented.)

Also add a jquery dot dot dot plugin to animate the waiting process
that the user has to go through when gathering the item details.