Implement add tracked item in UI
Once the item details are gathered, clicking Add will submit a POST
request to the server.  This request contains not just the URI, but
also the name, category, and price of the item.  A new tracked item and
price are created on the server side, and the UI takes the user back to
view all tracked items.