Undo AppFog configuration
Apparently AppFog has problems with contextify, or perhaps Python
libraries in general, and it's difficult to decipher the exact problem
because of the lack of access on the machine itself.  The error ends up
as:

"contextify.node: invalid ELF header"

Not really sure there, but a couple of searches online pulls up the
fact that others have hit this problem on AppFog.  The dependency comes
from scraper, which could be rewritten to use request directly… but
until then just disable AppFog since it's no longer in use.