Remove prices when tracked item deleted
Connect the price with the corresponding tracked item, so that when we
delete a tracked item, we also delete all the prices.  This is done via
a schema pre remove hook, which requires that we perform the remove on
the document, not the model.  So in the DELETE operation in the API,
find the tracked item, then perform the remove to trigger the pre
remove hook, which will delete all the associated prices.

Also, when we create new prices within a tracked item, we'll link it to
the tracked item.