Remove scraper lib, use request/cheerio directly
After seeing a couple out of memory exceptions, remove the dependency
on the scraper library (which uses jsdom, which might very well be the
cause of the memory exceptions) and instead use request directly along
with cheerio.  Cheerio gives us the benefit of getting jquery-like
searching without the requirement of jsdom and a small core subset of
jquery.  Hopefully this will fix any memory issues…

Also, add some timing to the log messages to keep track of how long
updating the prices takes.