Add footer, use it to display status
Add a footer to the UI, and within it, display status of the updater.

Create a Status model, and "offline" create a single status document
which will be updated each time the updater runs.  The status document
contains the last time updated and the duration of time the updater
takes to run.