Create TrackedFlight model, begin updateTrackedFlights
Create a mongo/mongoose model to represent a tracked flight.  This
model has information about the flight we’d like to track, along with
the prices for the flight (by day).

Begin to implement the updateTrackedFlights within the price updater
library.  This is partly from the logic written in the flights
controller, but also with an understanding that we’ll loop over all
tracked flights and find if the prices have changed for any days.

Also create some helpers in the flights API to test this out.