Better algorithm for limit of display tracked items
Move the logic to determine if we should display the tracked item to
the tracked-items-view.  In there, determine if the recent price change
had occurred, and if so, display the tracked item.  If not, skip
rendering of the tracked item.

Allow the user to click "display all" to force displaying of each
tracked item.

We still render the category heading, but not the tracked items within
the category.  By default, hide the category heading unless a tracked
item was rendered within that category.