Updates for Node v6
Update dependencies to support Node v6.

Apparently `zombie` has some issues with Node v6.  But the thing is,
`zombie` is only used in the tracked flights area of the application,
which was a prototype at best, and never really worked well.  To avoid
just deleting all the code, and I know that commenting out code is just
terrible, but in this case I’d like to do it.  I don’t want to have to
search through git history to see this code if I ever need it again.
And maybe one day `zombie` will update and work fine, and we can
uncomment all this code.  Maybe.  Until then, hack away!